work on db seed

+++ database/migrations/2017_04_03_124108_create_orders_table.php
@@ -15,7 +15,7 @@ class CreateOrdersTable extends Migration
     {
         Schema::create('orders', function (Blueprint $table) {
             $table->increments('id');
-            $table->float('price')->nullable();
+            $table->double('price')->nullable();
             $table->integer('user_id');
             $table->foreign('user_id')->references('users')->on('id')->onDelete('cascade');
             $table->integer('state_id');
